Вынеси в отдельные модули timer, checkResults, const
