+++ js/const.js
@@ -1,0 +1,16 @@
+export const MAX_QUESTION_AMOUNT = 10;
+
+export const ANSWER_VALUE = {
+  fast: 150,
+  right: 100,
+  slow: 50,
+};
+
+export const LIFE_VALUE = 50;
+
+export const answerStatus = {
+  fast: `fast`,
+  right: `right`,
+  slow: `slow`,
+  wrong: `wrong`,
+};
